interactive_db.py
# flake8: noqa
from crabber import app
from extensions import db
from models import (
    AccessToken,
    Bookmark,
    crabtag_table,
    Card,
    Crab,
    Crabtag,
    DeveloperKey,
    following_table,
    Like,
    Molt,
    Trophy,
    TrophyCase,
    ModLog,
    Notification,
    blocking_table,
    ImageDescription,
)
from sqlalchemy import desc, func, or_

app.app_context().push()
