pyproject.toml
[tool.poetry]
name = "crabber"
version = "0.1.0"
description = ""
authors = ["Jake Ledoux <contactjakeledoux@gmail.com>"]
license = "GNU General Public License v2.0"

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
Flask = "^2.0.2"
Flask-hCaptcha = "^0.5.4"
Flask-Limiter = "^1.4"
Flask-SQLAlchemy = "^2.5.1"
passlib = "^1.7.4"
crabatar = "^1.0.3"
geoip2 = "^4.4.0"
python-dotenv = "^0.19.1"
boto3 = "^1.19.7"
PyMySQL = "^1.0.2"
cryptography = "^35.0.0"
beautifulsoup4 = "^4.10.0"
webpreview = "^1.6.0"
gunicorn = "^20.1.0"
gevent = "^21.8.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
python-language-server = {extras = ["pycodestyle", "pydocstyle", "rope"], version = "^0.36.2"}
black = "^21.10b0"
pytest-pycodestyle = "^2.2.0"
pytest-flake8 = "^1.0.7"
pytest-pydocstyle = "^2.2.0"
pydocstyle = {extras = ["toml"], version = "^6.1.1"}

[tool.pytest.ini_options]
addopts = "--pycodestyle --pydocstyle --flake8 --ignore=scripts -W ignore::DeprecationWarning"

[tool.black]
# add black options when necessary

[tool.pydocstyle]
convention = "google"
add-ignore = [
    "D100",
    "D107",
    "D103",
    "D105"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
